orders

+++ lib/models/profilemodel.dart
@@ -29,7 +29,7 @@ class Data {
   String? phone;
   String? image;
   int? points;
-  int? credit;
+  double? credit;
   String? token;
 
   Data(
